+++ web/documentserver-example/nodejs/helpers/fileUtility.js
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -68,6 +68,7 @@ fileUtility.fileType = {
   word: 'word',
   cell: 'cell',
   slide: 'slide',
+  pdf: 'pdf',
 };
 
 fileUtility.getSuppotredExtensions = function getSuppotredExtensions() {