+++ web/documentserver-example/php/assets/js/jscript.js
@@ -1,6 +1,6 @@
 ﻿/**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -200,7 +200,7 @@ if (typeof jQuery != "undefined") {
 
         var fileName = jq("#hiddenFileName").val();
         var posExt = fileName.lastIndexOf('.');
-        posExt = 0 <= posExt ? fileName.substring(posExt).trim().toLowerCase() : '';
+        posExt = 0 <= posExt ? fileName.substring(posExt + 1).trim().toLowerCase() : '';
 
         if (EditedExtList.indexOf(posExt) != -1 || FillFormsExtList.indexOf(posExt) != -1) {
             jq("#beginEdit").removeClass("disable");