+++ web/documentserver-example/ruby/app/configuration/configuration.rb
@@ -1,5 +1,5 @@
 #
-# (c) Copyright Ascensio System SIA 2023
+# (c) Copyright Ascensio System SIA 2024
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,7 +31,7 @@ class ConfigurationManager
 
   sig { void }
   def initialize
-    @version = '1.7.0'
+    @version = '1.8.0'
   end
 
   sig { returns(T.nilable(URI::Generic)) }