+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Change Log
 
 - nodejs: delete all files
+- change insert image
 - nodejs: handling conversion -9 error
 - different goback for users
 - nodejs: converting function on index page
+
+## 1.8.0
+- nodejs: pdf, djvu, xps, oxps as pdf documentType
+- nodejs: filling pdf
 - version number to page meta
 - ar skin languages
 - sr-Latn-RS skin languages